refactor: 정산 조회와 송금 서비스를 기본 service 패키지로 통합
- query, transfer 하위 패키지 분리를 되돌리고 기존 service 패키지로 정리
- 아직 CQRS 단계가 아닌 현재 구조에 맞춰 서비스 책임을 단순화
